(PUP-7224) Improve Ruby class generation from TypeSet.
This commit improves how a `TypeSet` is generated as a ruby `module`
with respect to module nesting. It also adds the ability to override
the implementation registry using a provided regexp substitution. This
is handy when there's a need to generate and debug a module that already
exists and is loaded into the runtime.