(PUP-7224) Make it possible to register Pcore types in the static loader
The Pcore AST model must be bootstrapped before the Puppet Parser can
be used. Since the parser is used early in the boot process (to parse
types), the AST model must be present even earlier. This commit prepares
what's needed to add such "early needed" types to the static loader.