(PUP-8508) Add option to recognize backtick quoted strings to Puppet lexer
This commit adds an option hash to the puppet parser which is also
propagated to the puppet lexer. The parser will currently not look at this
hash but the lexer will recognize the key `:backtick_strings` as a boolean
option. The lexer will recognize strings quoted with backticks verbatim
when this option is enabled. In other words, what's between the backticks
will be lexed from start to end, linefeeds and all, as a string. Nothing,
not even the backtick character, can be escapded.

The concept of unescaped backtick qouted strings is borrowed from Golang
and its primary usage is expected to be when writing pspec-tests. Backtick
is an unusual character that most tests can do without and having a way of
expressing code without worrying about how escapes are handled is essential
when writing such tests.

The draback of not having an escape character is of course that a
a backticked string cannot contain a backtick chararacter. Such strings
must be expressed using heredoc or other types of quotes.